08_agentic_ai/rag_agent/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langgraph>=0.0.26
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
